requirements.txt
Django==5.1.3
mysqlclient>=2.0.3
pymysql>=1.0.3
openai>=0.27.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
pytest-django>=4.5.2
django-redis>=5.2.0
redis>=4.5.0
 